+++ internal/logging/archive_other.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package logging
+
+import "os"
+
+func createPrivateArchive(path string) (*os.File, error) {
+	return os.OpenFile(path, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0600)
+}